refactor: new context for issue view
