+++ apps/app/components/toast-alert/index.tsx
@@ -7,7 +7,7 @@ import {
   XCircleIcon,
   XMarkIcon,
 } from "@heroicons/react/24/outline";
-import useToast from "hooks/useToast";
+import useToast from "hooks/use-toast";
 // icons
 
 const ToastAlerts = () => {
