+++ apps/app/hooks/use-outside-click-detector.tsx
